[UPDATE] refactor steams read response and fix the error when target link is deleted.
