+++ Sources/KurrentDB/Core/Event/ReadEvent.swift
@@ -9,8 +9,23 @@ import Foundation
 import GRPCEncapsulates
 
 public struct ReadEvent: Sendable {
-    public internal(set) var recordedEvent: RecordedEvent
-    public internal(set) var linkedRecordedEvent: RecordedEvent?
+    
+    @available(*, deprecated, renamed: "link")
+    public var linkedRecordedEvent: RecordedEvent?{
+        get{
+            link
+        }
+    }
+    
+    @available(*, deprecated, renamed: "event")
+    public var recordedEvent: RecordedEvent{
+        get{
+            event
+        }
+    }
+    
+    public internal(set) var event: RecordedEvent
+    public internal(set) var link: RecordedEvent?
     public internal(set) var commitPosition: StreamPosition?
 
     public var noPosition: Bool {
@@ -18,24 +33,24 @@ public struct ReadEvent: Sendable {
     }
 
     package init(event: RecordedEvent, link: RecordedEvent? = nil, commitPosition: StreamPosition? = nil) {
-        recordedEvent = event
-        linkedRecordedEvent = link
+        self.event = event
+        self.link = link
         self.commitPosition = commitPosition
     }
 
     package init(message: EventStore_Client_Streams_ReadResp.ReadEvent) throws {
-        recordedEvent = try .init(message: message.event)
-        linkedRecordedEvent = try message.hasLink ? .init(message: message.link) : nil
-
-        if let position = message.position {
-            switch position {
-            case .noPosition:
-                commitPosition = nil
-            case let .commitPosition(commitPosition):
-                self.commitPosition = .at(commitPosition: commitPosition)
-            }
-        } else {
-            commitPosition = nil
+        let event: RecordedEvent = try .init(message: message.event)
+        let link: RecordedEvent? = try message.hasLink ? .init(message: message.link) : nil
+        
+        let commitPosition: StreamPosition? = switch message.position {
+        case .noPosition:
+            nil
+        case let .commitPosition(commitPosition):
+            .at(commitPosition: commitPosition)
+        case .none:
+            nil
         }
+        
+        self.init(event: event, link: link, commitPosition: commitPosition)
     }
 }
